pyproject.toml
[project]
name = "open-stream-deck"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "schema>=0.7.7",
    "streamdeck>=0.9.6",
    "websocket>=0.2.1",
    "websocket-client>=1.8.0",
]
